[types] rename currency module to community
