+++ packages/types/src/interfaces/community/definitions.ts
