+++ packages/types/src/interfaces/definitions.ts
@@ -1,5 +1,5 @@
 export { default as common } from './common/definitions';
 export { default as ceremony } from './ceremony/definitions';
 export { default as scheduler } from './scheduler/definitions';
-export { default as currency } from './currency/definitions';
+export { default as currency } from './community/definitions';
 export { default as worker } from './worker/definitions';
