+++ packages/types/src/interfaces/types.ts
@@ -4,5 +4,5 @@
 export * from './common/types';
 export * from './ceremony/types';
 export * from './scheduler/types';
-export * from './currency/types';
+export * from './community/types';
 export * from './worker/types';
